ogenti_platform/requirements.txt
# Ogenti Platform Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
stripe==10.12.0
python-multipart==0.0.9
jinja2==3.1.4
aiosqlite==0.20.0
